Removing interactivity to client-side component
